Comments Added / Cleaned Code

+++ Shared/Expense.cs
@@ -2,9 +2,10 @@
 {
     public class Expense
     {
+        /* Properties For Expenses Set */
         public int Id { get; set; }
         public string? Name { get; set; } = string.Empty;
-        public int Amount { get; set; }
+        public decimal Amount { get; set; }
         public ItemSet? ItemSet { get; set; }
         public int ItemSetId { get; set; }
     }
